Don't delete current object version on bad md5
Fixes #705

Previously, writing an object with an incorrect md5 would not only cause
that write to be garbage-collected, but would remove any existing object
at the key. This fix does not pro-actively schedule the incorrect object
for garbage-collection. This is because
`riak_cs_gc:gc_specific_manifests/5` expects a Riak Object, and we don't
have one as a return-value of `maybe_update_manifest_with_confirmation`.
Returning something other than `ok` from that, and related functions
would be a bit of a refactoring, and would provide information that most
callers don't need. This can be an optimization down the road, as the
bad object will still be garbage-collected once it's been seen that it
was never moved from `pending_delete` to `scheduled_delete`.