Ignore Owner with an empty ID in ACL set operations
Some clients (eg. Cyberduck) are a bit special and send Owner
elements in an ACL update with an empty ID inside when they are only
meant to be updating the ACL and not the owner at all. If we
get one of these, we should probably just ignore it rather than
crashing the request with a 500.