Fix handling of characters in XML generation
There are some codes where characters are represented by lists
those are generated from user-input binary (e.g. key names)
by binary_to_list/1.
It works for ASCII (or Latin-1) but not for unicode in general.

Handling of characters (binaries or lists of integers) of this fix are as follows:

- For binaries, we assume that they are encoded by UTF-8.
- For lists, we assume they are converted from UTF-8 encoded binaries
  by applying binary_to_list/1.
  This is tricky point but binary_to_list/1 treats binaries as if it is
  Latin-1 encoded, so we should turn them back by list_to_binary first.
  Then we can apply unicode:characters_to_binary safely.