Ignore one dialyzer error
I have not been able to solve this dialyzer error, so its going in the
new ignore file. I have not seen the error appear on a hipe-compiled
Erlang, but I'm not quite sure why that might be the case. I didn't
think hipe-compiled Erlang would differ in the Dialyzer code used.