Deprecate V2 Replication
As V2 replication is rarely used for CS, sometimes hard to maintain
and will be completely removed in future, with Riak CS 2.0 V2 replication
will not be supported any more. As there should be no code change
in Riak CS, CS is not sensitive to replication version.

Here is an upgrade guide:
http://docs.basho.com/riakee/latest/cookbooks/Multi-Data-Center-Replication-UpgradeV2toV3/