Filter out buckets in user record by last_action=deleted
This fixes a bug related to bucket name re-cycle,
storage calc wrongly includes deleted buckets.
An example of steps to reproduce is:

1. Alice creates a bucket foo
2. Alice deletes the bucket foo
3. Bob creates a bucket with the same name foo
4. Bob PUTs an object to the bucket
5. Then storage calc reports the bucket usage to both Alice and Bob

Condition: (A AND (B-1 OR B-2))

A. a user deletes a bucket and another user creates a bucket with the same name.

B-1. recycle of the bucket name is occured in leeway seconds (default: 1 day)
B-2. some nodes in cluster are down before another change to the user occures.