Handle socket error by catching badmatch
When a client cancelled or stopped uploading an object in the middle,
the socket for TCP connection returns error when mochiweb tried to
read payload. It had been throwing badmatch and that request was just
failing, leaving already uploaded chunks and writing manifests.

This commit adds a try-catch clause to catch the badmatch error thrown
by mochiweb and calls amendment code that moves the exact manifest to
GC bucket and marks it as scheduled_delete.

A test is included in regression_tests.erl, where a socket error is
emulated by client-side socket close in the middle of upload.

There is another possibility of cancelled upload in multipart upload,
where a part upload could be cancelled and left as it was, but it
can be collected if the multipart upload was aborted or completed.